+++ BaseLibS/Api/IRegressionFeatureRankingMethod.cs
@@ -3,7 +3,7 @@ using BaseLibS.Param;
 
 namespace BaseLibS.Api{
     public interface IRegressionFeatureRankingMethod: INamedListItem{
-		int[] Rank(BaseVector[] x, float[] y, Parameters param, IGroupDataProvider data, int nthreads);
+		int[] Rank(BaseVector[] x, double[] y, Parameters param, IGroupDataProvider data, int nthreads);
         Parameters GetParameters(IGroupDataProvider data);
     }
 }
